Move print_() functions to src/ui.rs
