+++ src/cmd.rs
@@ -1,4 +1,3 @@
-use colored::*;
 use regex::Regex;
 use structopt::StructOpt;
 
@@ -127,22 +126,6 @@ pub enum SubCommand {
 
     #[structopt(name = "upgrade-pip", about = "Upgrade pip in the virtualenv")]
     UpgradePip {},
-}
-
-pub fn print_error(description: &str) {
-    eprintln!("{}: {}", "Error".bold().red(), description);
-}
-
-pub fn print_warning(description: &str) {
-    eprintln!("{}: {}", "Warning".bold().yellow(), description);
-}
-
-pub fn print_info_1(message: &str) {
-    println!("{} {}", "::".blue(), message);
-}
-
-pub fn print_info_2(message: &str) {
-    println!("{} {}", "->".blue(), message);
 }
 
 // Make sure the `--python-version` option used in `dmenv lock`
