+++ src/operations/init.rs
@@ -1,5 +1,5 @@
-use crate::cmd::*;
 use crate::error::*;
+use crate::ui::*;
 use std::path::Path;
 
 #[derive(Debug)]
